This PR is the fourth step of #160175. It refactors the following:
CUDAKernel → CUTLASSKernel
CUDATemplateKernel → CUTLASSTemplateKernel
CUDATemplateCaller → CUTLASSTemplateCaller
CUDATemplateBuffer → CUTLASSTemplateBuffer



cc voznesenskym penguinwu EikanWang jgong5 Guobing-Chen XiaobingSuper zhuhaozhe blzheng wenzhe-nrv jiayisunx ipiszy chenyang78 kadeng muchulee8 amjames chauhang aakhundov coconutruben

[ghstack-poisoned]

+++ torch/csrc/distributed/autograd/context/container.cpp
@@ -52,13 +52,9 @@ DistAutogradContainer& DistAutogradContainer::init(int64_t worker_id) {
   }
 
   container.worker_id_ = static_cast<int16_t>(worker_id);
-  container.next_context_id_ = static_cast<int64_t>(worker_id)
-      << kAutoIncrementBits;
-  container.next_autograd_message_id_ = static_cast<int64_t>(worker_id)
-      << kAutoIncrementBits;
-  container.max_id_ =
-      (kAutoIncrementMask |
-       (static_cast<int64_t>(worker_id) << kAutoIncrementBits));
+  container.next_context_id_ = worker_id << kAutoIncrementBits;
+  container.next_autograd_message_id_ = worker_id << kAutoIncrementBits;
+  container.max_id_ = (kAutoIncrementMask | (worker_id << kAutoIncrementBits));
   container.initialized_ = true;
   return container;
 }
